refactor: improve pack list actions & error handlers

+++ src/utils/packs/create.js
@@ -14,5 +14,7 @@ module.exports = async (ctx, packToCreate) => {
 
   const createdPack = await ctx.getStickerSet(packToCreate.name);
   const stickerToDelete = createdPack.stickers.slice(-1)[0];
-  return ctx.deleteStickerFromSet(stickerToDelete.file_id);
+  if (stickerToDelete) {
+    return ctx.deleteStickerFromSet(stickerToDelete.file_id);
+  }
 };
